+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.1.6
 - #12 bin to scripts
+- #10 Rename images based on value of barcodes
 - #7 Travis build failing
 - #5 pep8 and code layout